% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircaPower.R
\name{CircaPower}
\alias{CircaPower}
\title{CircaPower}
\usage{
CircaPower(
  n = NULL,
  power = NULL,
  r = NULL,
  phi = 0,
  period = 24,
  cts = NULL,
  ct_estimation = TRUE,
  alpha = 0.05
)
}
\arguments{
\item{n}{Sample size.}

\item{power}{Statistical power.}

\item{r}{Intrinsic effect size. r=A/\eqn{\sigma}, where A is the amplitude of the sinusoidal curve: \eqn{y = A * sin(2\pi/period * (\phi + cts)) + M + \sigma}, and \eqn{\sigma} is the noise level of the independent Normal error term \eqn{N(0,\sigma^2)}.}

\item{phi}{Phase shift eqn{\phi} of the sinusoidal curve. Default is 0.}

\item{period}{Period of the sinusoidal curve. Default is 24.}

\item{cts}{Circadian times of the putative samples. If cts is NULL, evenly-spaced circadian time design will be used. If cts is not NULL,}

\item{ct_estimation}{If TRUE, we will calculate the sampling design factor using the observed circadian time; if FALSE, we will (1) perform kernel density estimation from the observed circadian time, (2) draw circadian times from the kernel density estimation, (3) calculate the sampling design factor using these drawn samples.}

\item{alpha}{Type I error control. Default is 0.05.}
}
\value{
A list of arguments (including the computed one) augmented.
}
\description{
Analytical circadian power calculation based on a cosiner model.
}
\details{
Exactly one of the parameters 'n','power', 'r' and 'alpha' must be passed as NULL, and that parameter is determined from the others.  Notice that alpha has non-NULL default so NULL must be explicitly passed if you want to compute it.
}
\examples{
CircaPower(n=NULL, power=0.8, r=1.5)

n = 100
cts = seq(0,24,length.out = n+1)[-1]
A = 1
sigma = 1
r = A/sigma
CircaPower(n, r=r, phi=0, period = 24, cts=cts, alpha = 0.05)
}
\author{
Wei Zong, Zhiguang Huo
}
