% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cts_Chen.R
\docType{data}
\name{cts_Chen}
\alias{cts_Chen}
\title{Brodmann’s Areas (BA) 11 and 47 Brain Aging Data Time of Death}
\format{
A numerical vector: time of death from Brodmann’s areas (BA) brain aging data.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE71620}
}
\usage{
data(cts_Chen)
}
\description{
Time of death vector from the human post-mortem brain transcriptomic study on Brodmann's area 11 and 47 in the prefrontal cortex.
}
\references{
Chen, C.-Y., Logan, R. W., Ma, T., Lewis, D. A., Tseng, G. C., Sibille, E., and McClung, C. A. (2016).Effects of aging on circadian patterns of gene expression in the human prefrontal cortex.Proceedings of the National Academy of Sciences, 113(1):206–211.
}
\keyword{datasets}
