% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cts_Ketchesin.R
\docType{data}
\name{cts_Ketchesin}
\alias{cts_Ketchesin}
\title{Dorsal and Ventral Straitum Data (control subjects) Time of Death}
\format{
A time of death vector.
}
\usage{
data(cts_Ketchesin)
}
\description{
Time of death vector from the human post-mortem brain transcriptomic study on the dorsal and ventral straitum.
}
\references{
Ketchesin, K. D., Zong, W., Hildebrand, M. A., Seney, M. L., Cahill, K. M., Scott, M. R., ... & McClung, C. A. (2021). Diurnal rhythms across the human dorsal and ventral striatum. Proceedings of the National Academy of Sciences, 118(2).
}
\keyword{datasets}
