% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cts_Chen}
\alias{cts_Chen}
\title{Human Postmortem Brain Data Time of Death}
\format{
Three numerical vectors:
\describe{
\item{cts_Chen}{Time of death from Brodmann’s areas (BA) brain aging data}
\item{cts_Seney}{Time of death from dorsal and ventral straitum data of control subjects.}
\item{cts_Ketchesin}{Time of death from dorsolateral prefrontal cortex data of control subjects.}
}#'
}
\usage{
data(cts_Chen)
data(cts_Seney)
data(cts_Ketchesin)
}
\description{
Time of death vectors from the human post-mortem brain transcriptomic studies.
}
\references{
Seney, M. L., Cahill, K., Enwright, J. F., Logan, R. W., Huo, Z., Zong, W., ... & McClung, C. A. (2019). Diurnal rhythms in gene expression in the prefrontal cortex in schizophrenia. Nature communications, 10(1), 1-11.
Ketchesin, K. D., Zong, W., Hildebrand, M. A., Seney, M. L., Cahill, K. M., Scott, M. R., ... & McClung, C. A. (2021). Diurnal rhythms across the human dorsal and ventral striatum. Proceedings of the National Academy of Sciences, 118(2).
Chen, C.-Y., Logan, R. W., Ma, T., Lewis, D. A., Tseng, G. C., Sibille, E., and McClung, C. A. (2016).Effects of aging on circadian patterns of gene expression in the human prefrontal cortex.Proceedings of the National Academy of Sciences, 113(1):206–211.
}
\keyword{datasets}
